+++ packages/web/src/styles/globals.css
@@ -7,7 +7,8 @@ body {
 }
 
 .input-reset {
-  all: unset;
+  border: none;
+  outline: none;
 }
 
 .chat-placeholder::placeholder {
